[fix]評価値のint型をdouble型にした

+++ NeoSolver/Evaluation/evaluation.h
@@ -7,7 +7,7 @@ class Evaluation
 {
 public:
     Evaluation();
-    std::vector<std::pair<int , Connect>> evaluation(procon::NeoExpandedPolygon const& frame , procon::NeoExpandedPolygon const& polygon);
+    std::vector<std::pair<double , Connect>> evaluation(procon::NeoExpandedPolygon const& frame , procon::NeoExpandedPolygon const& polygon);
 };
 
 #endif // EVALUATION_H
